Refactor frontend to feature-first structure
