+++ frontend/src/features/appointments/components/HistoryCard.jsx
