+++ frontend/src/features/auth/pages/RegisterPage.jsx
