+++ frontend/src/features/home/components/AboutSection.jsx
