+++ frontend/src/features/home/components/AuthSection.jsx
