+++ frontend/src/features/home/pages/HomePage.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { Link } from 'react-router-dom';
-import HomeFooter from '../components/home/HomeFooter';
-import { useSiteSettings } from '../SiteSettingsContext';
+import HomeFooter from '../components/HomeFooter';
+import { useSiteSettings } from '../../../shared/context/SiteSettingsContext';
 
 const testimonials = [
   {
