+++ frontend/src/features/patient/pages/PatientProfile.jsx
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useMemo, useState } from 'react';
-import { AuthContext } from '../AuthContext';
+import { AuthContext } from '../../auth/context/AuthContext';
 import { FaCalendarCheck, FaDownload, FaFileMedical, FaPrescriptionBottle } from 'react-icons/fa';
 
 function PatientProfile() {
