+++ frontend/src/features/services/pages/ServicesPage.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { useSiteSettings } from '../SiteSettingsContext';
+import { useSiteSettings } from '../../../shared/context/SiteSettingsContext';
 
 const apiBaseUrl = process.env.REACT_APP_API_URL;
